+++ CHANGELOG.md
@@ -12,11 +12,14 @@
 
 ### Added
 
+* Transactional Outbox Pattern with RabbitMQ
 * Support to delete container tree hierarchies
 * ODA trigger that inserts a dlm.data_item when an execution block is added.
 * dlm-archive stfp endpoint.
 * pg_sphere to local db build.
 * default dlm-archive storage endpoint.
+=======
+
 
 ## 2.1.0